requirements.txt
flask
torch
numpy
pillow
realesrgan
git+https://github.com/xinntao/Real-ESRGAN.git
